+++ dr-testing-framework/scripts/create_drt_dataverse_schema.py
@@ -173,7 +173,7 @@ COLUMNS = {
                 ]
             },
             "Format": "DateAndTime",
-            "DateTimeBehavior": {"Value": "UserLocal"},
+            "DateTimeBehavior": {"Value": "TimeZoneIndependent"},
         },
         # Actual recovery time in hours
         {